% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{bf_display}
\alias{bf_display}
\title{Choose display type for BF formatting.}
\usage{
bf_display(bf = NULL, display_type = "bf", k = 2)
}
\arguments{
\item{x}{A numeric vector containing one or more BF values.}
}
\value{
a formatted character string.
}
\description{
Choose display type for BF formatting.
}
\author{
Chuck Powell
}
