% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{BrownForsytheTest}
\alias{BrownForsytheTest}
\title{Brown-Forsythe Test for Homogeneity of Variance using median}
\usage{
BrownForsytheTest(formula, data)
}
\arguments{
\item{formula}{A fully crossed anova formula.}

\item{data}{A datafram containing the data.}
}
\value{
a table containing the results.
}
\description{
Brown-Forsythe Test for Homogeneity of Variance using median
}
\author{
J. Fox, Chuck Powell
}
\keyword{internal}
