% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{exponent}
\alias{exponent}
\title{Exponent of a number in scientific notation}
\usage{
exponent(x)
}
\arguments{
\item{x}{(required) numeric. A number.}
}
\value{
the exponent of the scientific notation representation of the number \code{x}
}
\description{
Returns the exponent of a number as it is written in scientific
  notation (powers of 10).
}
\references{
Thanks to Stackoverflow answer by Paul McMurdie \url{https://stackoverflow.com/a/25555105}
}
\author{
Tom Hopper
}
\keyword{internal}
