% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{justifyme}
\alias{justifyme}
\title{Justification for titles, subtitles and captions.}
\usage{
justifyme(x)
}
\arguments{
\item{x}{A numeric or character vector.}
}
\value{
a numeric value suitable for `ggplot2` `hjust` value.
}
\description{
Justification for titles, subtitles and captions.
}
\author{
Chuck Powell
}
\keyword{internal}
