% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{Mode}
\alias{Mode}
\title{Derive the modal value(s) for a set of data}
\usage{
Mode(x)
}
\arguments{
\item{x}{}
}
\value{
a list
}
\description{
This function takes a set of numeric data and returns one or mode values
that represent the mode point of the data
}
\examples{
Mode(mtcars$hp)
Mode(iris$Sepal.Length)

}
