% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{newgdp}
\alias{newgdp}
\title{Tufte dataset on Gross Domestic Product, 1970 and 1979}
\format{A data frame with 30 rows and 3 variables:
\describe{
  \item{Year}{character for 1970 and 1979}
  \item{Country}{factor country name}
  \item{GDP}{numeric a percentage of gross domestic product}

}}
\source{
Edward Tufte. \emph{Beautiful Evidence}. Graphics Press, 174-176.
}
\usage{
newgdp
}
\description{
Current receipts of fifteen national governments as a percentage of gross domestic product
}
\keyword{datasets}
