% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{not_a_factor}
\alias{not_a_factor}
\title{Test whether a vector or df column is not of type factor}
\usage{
not_a_factor(x)
}
\arguments{
\item{x}{an integer}
}
\description{
Test whether a vector or df column is not of type factor
}
\keyword{internal}
