% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{number_to_word}
\alias{number_to_word}
\title{Convert a vector of numbers to large-number word representation}
\usage{
number_to_word(x, nsmall = 0)
}
\arguments{
\item{x}{A vector of numbers to convert.}

\item{nsmall}{Optional. An integer number of digits to include to the right of the the leading digit}
}
\value{
A string representation of the number
}
\description{
Converts a vector of numbers to a character string approximation
  using the "short scale" version of large number names. e.g. 312e6 returns
  as '300 million.' Simultaneously returns a numeric representation of the
  approximation.
}
\author{
Tom Hopper
}
