% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotXtabs.R
\name{PlotXTabs}
\alias{PlotXTabs}
\title{Plot a Cross Tabulation of two variables using dplyr and ggplot2}
\usage{
PlotXTabs(dataframe, xwhich, ywhich, plottype = "side")
}
\arguments{
\item{dataframe}{an object that is of class dataframe}

\item{xwhich}{either a bare variable name that is valid in the dataframe or one or more column numbers}

\item{ywhich}{either a bare variable name that is valid in the dataframe or one or more column numbers}

\item{plottype}{one of three options "side", "stack" or "percent"}
}
\value{
one or more ggplots to the default graphics devise as well as advisory information in the console
}
\description{
Takes a dataframe and at least two variables as input, conducts a 
crosstabulation of the variables using dplyr removes NAs and then
plots the results as anyone of threee types of bar (column) graphs
using ggplot2.  The functions accepts either bare variable names or 
column numbers as input
}
\examples{
PlotXTabs(mtcars,am,vs)
PlotXTabs(mtcars,am,vs, "stack")
PlotXTabs(mtcars,am,vs, "percent")
PlotXTabs(mtcars,am,8, "side")
PlotXTabs(mtcars,8,am, "stack")
PlotXTabs(mtcars,am,c(8,10), "percent")
PlotXTabs(mtcars, c(10,8), am)
PlotXTabs(mtcars, c(2,9), c(10,8), "mispelled")

}
