% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pvalr}
\alias{pvalr}
\title{Internal function to format p values}
\usage{
pvalr(pvals, sig.limit = 0.001, digits = 3, html = FALSE)
}
\arguments{
\item{pvals}{a vector of p values}
}
\description{
Internal function to format p values
}
\keyword{internal}
