% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedist.R
\name{SeeDist}
\alias{SeeDist}
\title{See The Distribution}
\usage{
SeeDist(qqq, numbins = 0, whatvar = "Unspecified")
}
\arguments{
\item{qqq}{the data to be visualized}

\item{numbins}{the number of bins to use}

\item{whatvar}{additional information about the variable}
}
\value{
from 1 to 3 plots depending on what the user specifies
}
\description{
This function takes a vector of numeric data and returns one or more ggplot2
plots that help you visualize the data
}
\examples{
SeeDist(rnorm(100, mean=100, sd=20))
SeeDist(mtcars$hp, whatvar = "Horsepower")
SeeDist(iris$Sepal.Length, whatvar = "Sepal Length")

}
