% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{bf_display}
\alias{bf_display}
\title{Choose display type for BF formatting.}
\usage{
bf_display(bf = NULL, display_type = "bf", k = 2)
}
\arguments{
\item{bf}{A numeric vector containing one or more BF values.}

\item{display_type}{A string containing which option one of
"support", "logged", or "sensible".}

\item{k}{A numeric for the number of rounded digits.}
}
\value{
a formatted character string.
}
\description{
Choose display type for BF formatting.
}
\author{
Chuck Powell
}
