% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chaid_table.R
\name{chaid_table}
\alias{chaid_table}
\title{Produce CHAID results tables from a partykit CHAID model}
\usage{
chaid_table(chaidobject)
}
\arguments{
\item{chaidobject}{An object of type `constparty` or `party` which
was produced by `CHAID::chaid` see simple example below.}
}
\value{
A tibble containing the results.
}
\description{
Produce CHAID results tables from a partykit CHAID model
}
\examples{
library(CGPfunctions)
chaid_table(chaidUS)

}
\author{
Chuck Powell
}
