% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{Mode}
\alias{Mode}
\title{Derive the modal value(s) for a set of data}
\usage{
Mode(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a vector containing one or more modal values for the input vector
}
\description{
This function takes a vector and returns one or mode values
that represent the mode point of the data
}
\section{Warning}{

Be careful the function does some basic error checking but the return to
\code{Mode(NA)} is \code{NA} and a vector where the majority of entries
are \code{NA} is also NA
}

\examples{
Mode(sample(1:100,1000,replace=TRUE))
Mode(mtcars$hp)
Mode(iris$Sepal.Length)

}
