% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{newcancer}
\alias{newcancer}
\title{Tufte dataset on cancer survival rates}
\format{A data frame with 96 rows and 3 variables:
\describe{
  \item{Year}{ordered factor for the 5, 10, 15 and 20 year survival rates}
  \item{Type}{factor containing the name of the cancer type}
  \item{Survival}{numeric for this data a whole number corresponding to the percent survival rate}

}}
\source{
\url{https://www.edwardtufte.com/bboard/q-and-a-fetch-msg?msg_id=0003nk}
}
\usage{
newcancer
}
\description{
A dataset containing cancer survival rates for different types of cancer
over a 20 year period.
}
\keyword{datasets}
