% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newggslopegraph.R
\name{newggslopegraph}
\alias{newggslopegraph}
\title{Plot a Slopegraph a la Tufte using dplyr and ggplot2}
\usage{
newggslopegraph(dataframe, Times, Measurement, Grouping,
  Title = "No title given", SubTitle = "No subtitle given",
  Caption = "No caption given")
}
\arguments{
\item{dataframe}{a dataframe or an object that can be coerced to a dataframe. Basic error checking is performed.}

\item{Times}{a column inside the dataframe that will be plotted on the x axis. Traditionally this is some measure of time.  The function accepts a column of class ordered, factor or character.}

\item{Measurement}{a column inside the dataframe that will be plotted on the y axis. Traditionally this is some measure such as a percentage.  Currently the function accepts a column of type integer or numeric.}

\item{Grouping}{a column inside the dataframe that will be used to group and distinguish measurements.}

\item{Title}{Optionally the title to be displayed. Title = NULL will remove it entirely. Title = "" will provide and empty title but retain the sapcing.}

\item{SubTitle}{Optionally the sub-title to be displayed.  SubTitle = NULL will remove it entirely. SubTitle = "" will provide and empty title but retain the sapcing.}

\item{Caption}{Optionally the caption to be displayed. Caption = NULL will remove it entirely. Caption = "" will provide and empty title but retain the sapcing.}
}
\value{
a plot of type ggplot to the default plot device
}
\description{
Takes a dataframe as input, with three named columns being used to plot.
Makes the required adjustments to the ggplot2 parameters and returns the plot.
}
\examples{

newggslopegraph(newcancer, Year, Survival, Type)
newggslopegraph(newcancer, Year, Survival, Type, 
Title = "Estimates of Percent Survival Rates")


}
\seealso{
\code{\link{newcancer}}, \code{\link[car]{leveneTest}},
}
\author{
Chuck Powell
}
