% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seedist.R
\name{SeeDist}
\alias{SeeDist}
\title{See The Distribution}
\usage{
SeeDist(qqq, numbins = 0, whatvar = "Unspecified")
}
\arguments{
\item{qqq}{the data to be visualized}

\item{numbins}{the number of bins to use}

\item{whatvar}{additional information about the variable}
}
\value{
from 1 to 3 plots and some information in the console about mean, median and mode
}
\description{
See The Distribution
}
\examples{
SeeDist(mtcars$hp, whatvar = "Horsepower")
SeeDist(iris$Sepal.Length, whatvar = "Sepal Length")

}
